Rudimentary support for autocompletion of variables and method calls.
Also refactored the mechanism via which "goto" definitions are being
provided.

TODO: The completion `->`, if used outside of `{...}` breaks the Latte
parser (which is invoked, because the document changed and we want to scan
it to determine info about the variable for which we want to list methods),
because `>` looks like a closing of HTML tag, which "crashes"
the Latte parser, because it thinks we're exiting HTML region (in which
we may not even be).