DumbLatteParser now ignores errors while parsing Latte files.
It's better for the extension that the parser should try very hard not
to crash completely during parsing and to provide at least _some_
information even when there's a syntax (or other) error. So now we do that.

This change allows for parsing double Latte tags, for example when
used in `<script n:syntax="double"> const x = {{$var}}; </script>`.

The `n:syntax` tag actually doesn't do anything for us, but we just don't
crash when we encounter `{{` or `}}` _(because ordinarily that would
be interpreted as entering the Latte region-type twice, which would be
considered a syntax error.)_